ファイル分割の見直し / 関数からimplへの移行
