+++ server/app/Cargo.toml
@@ -15,3 +15,12 @@ once_cell = "1.19.0"
 uuid = { version = "1.8.0", features = ["v7"] }
 tower-http = {version = "0.5.2", features = ["fs"]}
 bcrypt = "0.15.1"
+async-session = "3.0.0"
+anyhow = "1.0.86"
+serde = "1.0.204"
+
+[dependencies.async-sqlx-session]
+git = "https://github.com/maxcountryman/async-sqlx-session.git"
+default-features = false
+branch = "sqlx-0.7"
+features = ["mysql"]
